Move net deposit calculation and remaining debt calculation into each corresponding tab, outside report panel; clear favorite map each time checked item is changed so to load latest data
