+++ src/pers/bocky/finance/view/impl/ConsumePanel.java
@@ -207,13 +207,11 @@ public class ConsumePanel extends JPanel implements WillBeInMainTabbed{
 		dp = new DateField();
 		
 		JButton calBtn = new JButton("求和");
-		JLabel sumLabel = new JLabel("");
-		sumLabel.setForeground(Color.RED);
 		calBtn.addActionListener(new ActionListener() {
 			
 			@Override
 			public void actionPerformed(ActionEvent e) {
-				sumLabel.setText(NumberFormat.getInstance().format(sumAmount()));
+				JOptionPane.showMessageDialog(ConsumePanel.this, NumberFormat.getNumberInstance().format(sumAmount()), "本页账目总和", JOptionPane.INFORMATION_MESSAGE);
 			}
 		});
 		
@@ -252,8 +250,6 @@ public class ConsumePanel extends JPanel implements WillBeInMainTabbed{
 		inputPanel.add(deleteBtn);
 		
 		inputPanel.add(calBtn);
-		inputPanel.add(sumLabel);
-		
 		
 		panel.add(inputPanel);
